+++ webpackVersion/scripts/content/database.js
@@ -1,0 +1,17 @@
+module.exports = {
+  title: 'Database',
+  contents: [
+    {
+      title: 'MySQL开发Note',
+      date: '2013-09-22',
+      url: 'static/notes/mysql/notes.html',
+      des: 'MySQL开发Note'
+    },
+    {
+      title: 'MongoDB开发Note',
+      date: '2013-09-22',
+      url: 'static/notes/mongodb/notes.html',
+      des: 'MongoDB开发Note'
+    }
+  ]
+};