+++ webpackVersion/scripts/content/opus.js
@@ -1,0 +1,11 @@
+module.exports = {
+	title: 'My Opus',
+	contents: [
+		{
+			title: '专注力测试',
+			date: '2016-10-01',
+			url: 'static/notes/opus/concentration/index.html',
+			des: '专注力测试'
+		}
+	]
+};