+++ webpackVersion/scripts/content/webpack.js
@@ -1,16 +1,21 @@
 module.exports= {
-	title: 'Webpack',
-	contents: [
-		{
-			title: 'Webpack introduction',
-			date: '2016-03-22',
-			url: 'static/notes/tools/webpack-introduction.html',
-		},
-		{
-			title: 'Webpack概述',
-			date: '2016-02-26',
-			url: 'static/notes/tools/webpack.html',
-			des: 'Webpack概述'
-		}
-	]
+  title: 'Webpack',
+  contents: [
+    {
+      title: 'Webpack Loaders',
+      date: '2016-10-30',
+      url: 'static/notes/tools/webpack-loaders.html',
+    },
+    {
+      title: 'Webpack introduction',
+      date: '2016-03-22',
+      url: 'static/notes/tools/webpack-introduction.html',
+    },
+    {
+      title: 'Webpack概述',
+      date: '2016-02-26',
+      url: 'static/notes/tools/webpack.html',
+      des: 'Webpack概述'
+    }
+  ]
 };